se elimina cartManager

+++ static/js/adminPanel.js
@@ -11,47 +11,51 @@ fetch("http://localhost:8080/profile/current")
 
 //DELETE USER
 const btnDelete = document.getElementById("btnDelete");
-btnDelete.addEventListener("click", () => {
-    const id = document.getElementById("id_delete").value;
-    if (id != "") {
-        fetch(`http://localhost:8080/api/users/${id}`, {
-            method: "DELETE",
-        })
-            .then((res) => res.json())
-            .then((res) => {
-                let message = res.message;
-                if (res.description) {
-                    message = res.description;
-                }
-                Swal.fire({
-                    title: "Resultado:",
-                    text: `${message}`,
+if (btnDelete) {
+    btnDelete.addEventListener("click", () => {
+        const id = document.getElementById("id_delete").value;
+        if (id != "") {
+            fetch(`http://localhost:8080/api/users/${id}`, {
+                method: "DELETE",
+            })
+                .then((res) => res.json())
+                .then((res) => {
+                    let message = res.message;
+                    if (res.description) {
+                        message = res.description;
+                    }
+                    Swal.fire({
+                        title: "Resultado:",
+                        text: `${message}`,
+                    });
+                    socket.emit("updateUsers");
                 });
-                socket.emit("updateUsers");
-            });
-    }
-});
+        }
+    });
+}
 
 //Change role
 const btnChangeRole = document.getElementById("btnChangeRole");
-btnChangeRole.addEventListener("click", () => {
-    const id = document.getElementById("id_role").value;
-    if (id != "") {
-        fetch(`http://localhost:8080/api/users/premium/${id}`)
-            .then((res) => res.json())
-            .then((res) => {
-                let message = res.message;
-                if (res.description) {
-                    message = res.description;
-                }
-                Swal.fire({
-                    title: "Resultado:",
-                    text: `${message}`,
+if (btnChangeRole) {
+    btnChangeRole.addEventListener("click", () => {
+        const id = document.getElementById("id_role").value;
+        if (id != "") {
+            fetch(`http://localhost:8080/api/users/premium/${id}`)
+                .then((res) => res.json())
+                .then((res) => {
+                    let message = res.message;
+                    if (res.description) {
+                        message = res.description;
+                    }
+                    Swal.fire({
+                        title: "Resultado:",
+                        text: `${message}`,
+                    });
+                    socket.emit("updateUsers");
                 });
-                socket.emit("updateUsers");
-            });
-    }
-});
+        }
+    });
+}
 
 //Renderizar
